Merged in jc-peer-dependency (pull request #2)
DLS Version 2.0.0

This is a major release of the DLS, primarily focused on @material-ui/core being added as a peer dependency so that most of the *Base components can be removed in favor of developers importing from @material-ui/core directly.

Check the README.md for updated installation and usage instructions.